+++ pymongo/encryption_options.py
@@ -18,7 +18,7 @@
 """
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any, Mapping, Optional
+from typing import TYPE_CHECKING, Any, Mapping, Optional, TypedDict
 
 from pymongo.uri_parser_shared import _parse_kms_tls_options
 
@@ -295,3 +295,85 @@ class RangeOpts:
             if v is not None:
                 doc[k] = v
         return doc
+
+
+class TextOpts:
+    """**BETA** Options to configure encrypted queries using the text algorithm.
+
+    TextOpts is currently unstable API and subject to backwards breaking changes."""
+
+    def __init__(
+        self,
+        substring: Optional[SubstringOpts] = None,
+        prefix: Optional[PrefixOpts] = None,
+        suffix: Optional[SuffixOpts] = None,
+        case_sensitive: Optional[bool] = None,
+        diacritic_sensitive: Optional[bool] = None,
+    ) -> None:
+        """Options to configure encrypted queries using the text algorithm.
+
+        :param substring: Further options to support substring queries.
+        :param prefix: Further options to support prefix queries.
+        :param suffix: Further options to support suffix queries.
+        :param case_sensitive: Whether text indexes for this field are case sensitive.
+        :param diacritic_sensitive: Whether text indexes for this field are diacritic sensitive.
+
+        .. versionadded:: 4.15
+        """
+        self.substring = substring
+        self.prefix = prefix
+        self.suffix = suffix
+        self.case_sensitive = case_sensitive
+        self.diacritic_sensitive = diacritic_sensitive
+
+    @property
+    def document(self) -> dict[str, Any]:
+        doc = {}
+        for k, v in [
+            ("substring", self.substring),
+            ("prefix", self.prefix),
+            ("suffix", self.suffix),
+            ("caseSensitive", self.case_sensitive),
+            ("diacriticSensitive", self.diacritic_sensitive),
+        ]:
+            if v is not None:
+                doc[k] = v
+        return doc
+
+
+class SubstringOpts(TypedDict):
+    """**BETA** Options for substring text queries.
+
+    SubstringOpts is currently unstable API and subject to backwards breaking changes.
+    """
+
+    # strMaxLength is the maximum allowed length to insert. Inserting longer strings will error.
+    strMaxLength: int
+    # strMinQueryLength is the minimum allowed query length. Querying with a shorter string will error.
+    strMinQueryLength: int
+    # strMaxQueryLength is the maximum allowed query length. Querying with a longer string will error.
+    strMaxQueryLength: int
+
+
+class PrefixOpts(TypedDict):
+    """**BETA** Options for prefix text queries.
+
+    PrefixOpts is currently unstable API and subject to backwards breaking changes.
+    """
+
+    # strMinQueryLength is the minimum allowed query length. Querying with a shorter string will error.
+    strMinQueryLength: int
+    # strMaxQueryLength is the maximum allowed query length. Querying with a longer string will error.
+    strMaxQueryLength: int
+
+
+class SuffixOpts(TypedDict):
+    """**BETA** Options for suffix text queries.
+
+    SuffixOpts is currently unstable API and subject to backwards breaking changes.
+    """
+
+    # strMinQueryLength is the minimum allowed query length. Querying with a shorter string will error.
+    strMinQueryLength: int
+    # strMaxQueryLength is the maximum allowed query length. Querying with a longer string will error.
+    strMaxQueryLength: int